pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
  { name = "Charlie Bini", email = "5003326+cbini@users.noreply.github.com" },
]
license-expression = "MIT"
dependencies = [
  "njdoe>=0.1.0",
  "google-cloud-storage>=2.2.1",
  "oauthlib>=3.2.0",
  "requests-oauthlib>=1.3.1",
]
requires-python = ">=3.9"

[project.urls]
Homepage = ""

[project.optional-dependencies]
[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = []

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
_.env_file = ".env"
bg-check = { call = "njdoe_sync.background_checks:main" }
post_lock = { shell = "pdm export -f requirements > requirements.txt" }
